🔥 Componentized About Us page
Restructured About Us page into components that can be shared across different pages" #9
