+++ client/src/components/PageHeader.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+const styles = {
+  header: {
+    paddingLeft: '3rem',
+    paddingTop: '3rem',
+    marginTop: '5rem',
+    fontFamily: 'Roboto',
+    fontSize: '3rem',
+    fontWeight: 'bold'
+  }
+}
+
+const PageHeader = (props) => {
+  return (
+    <div style={styles.header}>
+      <h1>{props.title}</h1>
+    </div>
+  )
+}
+
+export default PageHeader;
